requirements.txt
Django == 3.2
djangorestframework == 3.12.4
django-cors-headers == 3.7.0
mysqlclient == 2.0.3
mod-wsgi == 4.7.1